# backend/docs/openapi.yaml
openapi: 3.0.0
paths:
  /api/health:
    get:
      operationId: HealthController_check
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: ok
      summary: Liveness probe
      tags:
        - Health
  /api/users:
    post:
      operationId: UserController_register
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserDto'
      responses:
        '201':
          description: User registered successfully
        '400':
          description: ''
          content:
            application/json:
              schema:
                properties:
                  type:
                    type: string
                    example: https://my-api-doc.dev/problems/validation-error
                  title:
                    type: string
                    example: Validation Error
                  status:
                    type: number
                    example: 400
                  detail:
                    type: string
                    example: One or more fields failed validation.
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        field:
                          type: string
                          example: email
                        message:
                          type: string
                          example: email must be an email
        '409':
          description: ''
          content:
            application/json:
              schema:
                properties:
                  type:
                    type: string
                    example: https://my-api-doc.dev/problems/user-already-exists
                  title:
                    type: string
                    example: User Already Exists
                  status:
                    type: number
                    example: 409
                  detail:
                    type: string
                    example: User already exists with email john.doe@example.com
                  email:
                    type: string
                    example: john.doe@example.com
      summary: Register a new user
      tags:
        - Users
  /api/users/me:
    get:
      operationId: UserController_profile
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                properties:
                  id:
                    type: string
                    example: 550e8400-e29b-41d4-a716-446655440000
                  email:
                    type: string
                    example: john.doe@example.com
                  firstName:
                    type: string
                    example: John
                  lastName:
                    type: string
                    example: Doe
        '401':
          description: Missing or invalid JWT token
          content:
            application/json:
              schema:
                properties:
                  type:
                    type: string
                    example: about:blank
                  title:
                    type: string
                    example: Unauthorized
                  status:
                    type: number
                    example: 401
        '404':
          description: ''
          content:
            application/json:
              schema:
                properties:
                  type:
                    type: string
                    example: https://my-api-doc.dev/problems/entity-not-found
                  title:
                    type: string
                    example: Entity not found
                  status:
                    type: number
                    example: 404
                  detail:
                    type: string
                    example: >-
                      Entity not found with id
                      550e8400-e29b-41d4-a716-446655440000
                  entityId:
                    type: string
                    example: 550e8400-e29b-41d4-a716-446655440000
      security:
        - bearer: []
      summary: Get current user profile
      tags:
        - Users
  /api/auth/login:
    post:
      operationId: AuthController_login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginUserDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                properties:
                  accessToken:
                    type: string
        '400':
          description: ''
          content:
            application/json:
              schema:
                properties:
                  type:
                    type: string
                    example: https://my-api-doc.dev/problems/validation-error
                  title:
                    type: string
                    example: Validation Error
                  status:
                    type: number
                    example: 400
                  detail:
                    type: string
                    example: One or more fields failed validation.
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        field:
                          type: string
                          example: email
                        message:
                          type: string
                          example: email must be an email
        '401':
          description: ''
          content:
            application/json:
              schema:
                properties:
                  type:
                    type: string
                    example: https://my-api-doc.dev/problems/invalid-credentials
                  title:
                    type: string
                    example: Invalid Credentials
                  status:
                    type: number
                    example: 401
                  detail:
                    type: string
                    example: Invalid credentials provided.
      summary: Login with email and password
      tags:
        - Auth
  /api/credit/increase:
    post:
      operationId: CreditController_increase
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncreaseCreditDto'
      responses:
        '204':
          description: Credit increased successfully
        '400':
          description: ''
          content:
            application/json:
              schema:
                properties:
                  type:
                    type: string
                    example: https://my-api-doc.dev/problems/validation-error
                  title:
                    type: string
                    example: Validation Error
                  status:
                    type: number
                    example: 400
                  detail:
                    type: string
                    example: One or more fields failed validation.
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        field:
                          type: string
                          example: email
                        message:
                          type: string
                          example: email must be an email
        '401':
          description: Missing or invalid JWT token
          content:
            application/json:
              schema:
                properties:
                  type:
                    type: string
                    example: about:blank
                  title:
                    type: string
                    example: Unauthorized
                  status:
                    type: number
                    example: 401
        '409':
          description: ''
          content:
            application/json:
              schema:
                properties:
                  type:
                    type: string
                    example: https://my-api-doc.dev/problems/concurrent-modification
                  title:
                    type: string
                    example: Concurrent Modification
                  status:
                    type: number
                    example: 409
                  detail:
                    type: string
                    example: >-
                      Could not update the wallet: too many concurrent
                      modifications.
      security:
        - bearer: []
      summary: Increase the current user's account credit
      tags:
        - Credit
  /api/credit:
    get:
      operationId: CreditController_balance
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                properties:
                  amount:
                    type: number
                    example: 50000
        '401':
          description: Missing or invalid JWT token
          content:
            application/json:
              schema:
                properties:
                  type:
                    type: string
                    example: about:blank
                  title:
                    type: string
                    example: Unauthorized
                  status:
                    type: number
                    example: 401
      security:
        - bearer: []
      summary: Get the current user's account credit
      tags:
        - Credit
info:
  title: API
  description: ''
  version: '1.0'
  contact: {}
tags: []
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    RegisterUserDto:
      type: object
      properties:
        email:
          type: string
          format: email
          example: john.doe@example.com
        password:
          type: string
          minLength: 12
          example: password12345
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
      required:
        - email
        - password
        - firstName
        - lastName
    LoginUserDto:
      type: object
      properties:
        email:
          type: string
          format: email
          example: john.doe@example.com
        password:
          type: string
          example: password123
      required:
        - email
        - password
    IncreaseCreditDto:
      type: object
      properties:
        amount:
          type: number
          minimum: 1
          example: 50000
          description: Amount to add, in Rials.
      required:
        - amount